fixed git issue & added cors

+++ src/core/gcp/file_manager.py
@@ -1,10 +1,14 @@
 from typing import Dict,List
 from core.settings import app_settings
 from core.gcp import git
+from models.abm import Abm
+
 import yaml
 import os
 import shutil
 import logging
+
+
 
 def create_selector(selector_name: str, match_labels: Dict[str, str]):
     """ This function creates a cluster config selector and add its to the git repo"""
@@ -33,14 +37,15 @@ def create_selector(selector_name: str, match_labels: Dict[str, str]):
         folder_path = f"{app_settings.save_file_directory}selectors"
 
         # Build folder if not present
-        if not os.path.exists (folder_path):
+        if not os.path.exists(folder_path):
             os.makedirs(folder_path)
 
         save_file = f"{folder_path}/{selector_name}.yaml"
 
         # Check if selector already exists
         if os.path.exists(save_file):
-            return ""
+            logging.debug(f"Existing file found: {save_file}")
+            return save_file
 
         # Write file locally
         with open(save_file, 'w') as yaml_file:
@@ -77,10 +82,15 @@ def create_cluster(cluster_name: str, labels: Dict[str,str]):
         folder_path = f"{app_settings.save_file_directory}setup"
 
         # Build folder if not present
-        if not os.path.exists (folder_path):
+        if not os.path.exists(folder_path):
             os.makedirs(folder_path)
 
         save_file = f"{folder_path}/{cluster_name}.yaml"
+
+        # Check if cluster already exists
+        if os.path.exists(save_file):
+            logging.debug(f"Existing file found: {save_file}")
+            return save_file
 
         # Write file locally
         with open(save_file, 'w') as yaml_file:
@@ -123,10 +133,16 @@ def create_repo_sync(cluster_name: str, selector_name: str ):
         folder_path = f"{app_settings.save_file_directory}{cluster_name}"
 
         # Build folder if not present
-        if not os.path.exists (folder_path):
+        if not os.path.exists(folder_path):
             os.makedirs(folder_path)
 
         save_file = f"{folder_path}/repo-sync.yaml"
+
+
+        # Check if repo-sync already exists
+        if os.path.exists(save_file):
+            logging.debug(f"Existing file found: {save_file}")
+            return save_file
 
         # Write file locally
         with open(save_file, 'w') as yaml_file:
@@ -173,10 +189,16 @@ def create_role_binding(cluster_name: str, selector_name: str):
         folder_path = f"{app_settings.save_file_directory}{cluster_name}"
 
         # Build folder if not present
-        if not os.path.exists (folder_path):
+        if not os.path.exists(folder_path):
             os.makedirs(folder_path)
 
         save_file = f"{folder_path}/repo-rbac.yaml"
+
+
+        # Check if rbac already exists
+        if os.path.exists(save_file):
+            logging.debug(f"Existing file found: {save_file}")
+            return save_file
 
         # Write file locally
         with open(save_file, 'w') as yaml_file:
@@ -191,41 +213,88 @@ def create_role_binding(cluster_name: str, selector_name: str):
     # Return path to file
     return save_file
 
-def rebuild_clusters(cluster_name: str, cluster_labels: Dict[str,str], git_push: bool = True) -> list[str]:
+def create_repo_file(file_name: str, file_contents: str, basefolder: str = "default") -> str:
+    """ This function creates local file"""
+
+    save_file = ""
+
+    try:
+        save_file = f"{app_settings.save_file_directory}{basefolder}/{file_name}"
+        # Check if selector already exists
+        if os.path.exists(save_file):
+            logging.debug(f"Existing file found: {save_file}")
+            return ""
+
+        folder_path = os.path.dirname(save_file)
+
+        # Check if file already exists
+        if os.path.exists(save_file):
+            logging.debug(f"Existing file found: {save_file}")
+            return save_file
+
+        # Build folder if not present
+        if not os.path.exists(folder_path):
+            os.makedirs(folder_path)
+
+        # Write file locally
+        out_file = open(save_file, 'w')
+        out_file.write(file_contents)
+        out_file.close()
+
+    except Exception as e:
+        logging.error(e)
+        print(e)
+
+    return save_file    
+
+def rebuild_clusters(cluster_list: List[Abm]) -> List[str]:
     """ This function builder creates objects per cluster"""
 
+    # Source File list
     file_list=[]
+    # Return list
+    git_added_files = []
 
     try: 
-        # First create cluster object
-        file_list.append(create_cluster(cluster_name=cluster_name, labels=cluster_labels))
-        # Create Repo Sync File
-        file_list.append(create_repo_sync(cluster_name=cluster_name,selector_name=f"{cluster_name}-sel"))
-        # Create RBAC roles
-        file_list.append(create_role_binding(cluster_name=cluster_name, selector_name=f"{cluster_name}-sel"))
+        # Create All Cluster Selector
+        file_list.append(create_selector(selector_name=f"all-sel", match_labels={"loc":'*'}))
+        # Create Default apps
+        git_file_to_create = git.get_file_contents()
+        for key,value in git_file_to_create.items():
+            file_list.append(create_repo_file(file_name=key, file_contents=value))
 
-        # Then create selectors
-        ## Default selector
-        file_list.append(create_selector(selector_name=f"{cluster_name}-sel", match_labels={"loc":cluster_labels["loc"]}))
-        ## One for each label
-        for key, value in cluster_labels.items():
-            if key != "loc":
-                file_list.append(create_selector(selector_name=f"{key}-{value}-sel", match_labels={key:value}))
+        # Add cluster specific features
+        for aCluster in cluster_list:
 
-        # Cleanup file list
-        while("" in file_list):
-            file_list.remove("")
-        # If sync to git, sync to git
-        if git_push:
-            print(file_list)
-            git.add_file_to_branch(file_list=file_list)
-            cleanup_local_folder()
+            # Cluster Object
+            file_list.append(create_cluster(cluster_name=aCluster.name, labels=aCluster.labels))
+            # Repo Sync File
+            file_list.append(create_repo_sync(cluster_name=aCluster.name,selector_name=f"{aCluster.name}-sel"))
+            # RBAC roles
+            file_list.append(create_role_binding(cluster_name=aCluster.name, selector_name=f"{aCluster.name}-sel"))
+
+            # location selector
+            file_list.append(create_selector(selector_name=f"{aCluster.name}-sel", match_labels={"loc":aCluster.labels["loc"]}))
+            
+            # per-label selector
+            for key, value in aCluster.labels.items():
+                if key != "loc":
+                    file_list.append(create_selector(selector_name=f"{key}-{value}-sel", match_labels={key:value}))
+
+            # Cleanup file list
+            while("" in file_list):
+                file_list.remove("")
+        
+        # Finally sync to git
+        git_added_files = git.add_file_to_branch(file_list=file_list)
+        cleanup_local_folder()
+
     except Exception as e:
         logging.error(e)
         print(e)
      
 
-    return file_list
+    return git_added_files
 
 def cleanup_local_folder() -> bool:
     """ This function removes local folder"""
